Tweaked optimization.jl to replicate Soldatenko
Added an optional constraint on the final state, which allows us to almost exactly replicate the main results of the Soldatenko and Yusupov 2017 paper on optimal climate engineering.

I am hoping to start a library of examples where we replicate the results of other papers (esp. IAMs) with very minimal changes to the default configuration of the model.