Successfully replicated key Belaia results
This was a very valuable exercise which revealed a fundamental defiency in our model formulation as well as a bug (in the slow response to forcing).

For a while I was unable to reproduce Belaia's results and found (with her mitigation and solar geoengineering scaling costs) that solar geoengineering was far and above the most effective / net-beneficial control. The optimal solution was simply to not mitigate any emissions and just ramp up as much geoengineering as necessary in 2150-2250. After reading through Belaia's paper and the DICE documentation, I realized that while mitigation DOES NOT scale with GWP (it is a fixed
inflation-controlled price per tCO2), geoengineering's cost DOES scale with GWP. The cost of geoengineering is equated to the corresponding damages of an equal amount of CO2 forcing. Since GWP increases at about 2% per year (roughly equal to exp(t/50 years)), geoengineering becomes disadvantageous relative to mitigation in the future simply because the damages at a fixed level of geoengineering radiative forcing increase!

While trying to replicate Belaia's results (DICE documentation is... lacking), I stumbled upon a great Geoffrey 2013 paper where they estimate the free parameters (Cd, kappa, B) in the two-box energy balance model. Our default configuration now uses their values.