Avoid costs, combinatorics, two-box bug fix
I've added some calculations of control costs avoided by re-evaluate policies based on outcomes / new information.

I added in simulations with all combinations of the four controls {M,A,R,G} to estimate the benefits of considering a complete portfolio of climate controls, which now populate Table 1 in the correspoding version of the manuscript.

I fixed a few small bugs in the temperature kernel of the two-box model (only in the JuMP implementations; the diagnostic equations were fine), which don't seem to change the results much.

I also improved the replication of Belaia's Figure 4 by increasing the fraction of emissions that stay in the atmosphere. Before this change, our replication of their baseline emissions was quite good but it did not translate into good agreement in the radiative forcing (we underestimated it by ~20%). Now, we have quite good agreement between our RF and theirs. I've also decreases the maximum allowable deployment rate from 1/50 to 1/75 in order to better match their control deployment curves.

I also put in a few small quality-of-life changes, such as making the optimization algorithm print its raw status every time it is run, since I occasionally was confused by output that corresponded to an infeasible optimization.