Updated A&R costs; GtCO2 bugfix; updated q(t)
I updated A&R costs, decreasing R costs by ~30% (thanks to bugfix in GtCO2 to ppm conversions and the value of q0 used in the calculation of the cost estimate, which was hard coded to the wrong value) and A costs by 50% (no longer doubling it to account for unquantifiable adaptation sectors).

I also updated q(t), such that initial CO2e emissions are more consistent with present-day emissions. Now starts at 7.5 ppm/year in the baseline, quadrupling to 30 ppm/year by 2100.